+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   *  filters out dates with single polarization 
   *  allows users to select which dual polarization to use (`HH+HV` vs. `VV+VH`)
 * Baseline curation matches dist-s1.
+* Add models for v0 and v1 that use 32 x 32 input size.
+* Dynamically update input size for inference using model attribute.
+* Expose fill value that was reminant of old logit transformation and is now explicit
 
 
 ## [1.0.3] - 2025-08-25